#33 wip split shared utils to format libs
