+++ src/features/currency/components/Converter/Converter.tsx
@@ -1,6 +1,6 @@
 import { useState, type FunctionComponent } from 'react';
 import { LineField } from '../../../../components/LineField';
-import { formatToConvertNumber } from '../../../../utils';
+import { formatToConvertNumber } from '../../lib/format';
 import type { ConverterProps } from './Converter.types';
 import styles from './Converter.module.css';
