+++ src/features/currency/lib/format.ts
@@ -1,9 +1,7 @@
 import {
-  EXPLORE_FRACTION_DIGITS,
   CONVERT_FRACTION_DIGITS,
-} from './features/currency/constants';
-
-// @todo: This file must be split into separate shared libs
+  EXPLORE_FRACTION_DIGITS,
+} from '../constants';
 
 const exploreFormatter = new Intl.NumberFormat(
   navigator?.languages || ['en'],
@@ -20,16 +18,6 @@ const convertFormatter = new Intl.NumberFormat('en-US', {
   roundingMode: 'halfEven',
   useGrouping: false,
 });
-
-export const formatToIsodate = (dateInit: string | number): string => {
-  return new Date(dateInit).toISOString().slice(0, 10);
-};
-
-export const formatToIsodatetime = (
-  dateInit: string | number
-): string => {
-  return new Date(dateInit).toJSON().slice(0, 16).replace('T', ' ');
-};
 
 export const formatToExploreNumber = (numberInit: number): string => {
   return exploreFormatter.format(numberInit);
